+++ magda-web-client/src/config.js
@@ -22,7 +22,8 @@ const serverConfig: {
     previewMapBaseUrl?: string,
     registryApiBaseUrl?: string,
     searchApiBaseUrl?: string,
-    correspondenceApiBaseUrl?: string
+    correspondenceApiBaseUrl?: string,
+    gapiIds?: Array<string>
 } =
     window.magda_server_config || {};
 
@@ -156,5 +157,6 @@ export const config = {
     },
     fallbackUrl: serverConfig.fallbackUrl,
     datasetSearchSuggestionScoreThreshold:
-        serverConfig.datasetSearchSuggestionScoreThreshold || 65
+        serverConfig.datasetSearchSuggestionScoreThreshold || 65,
+    gapiIds: serverConfig.gapiIds || []
 };