+++ magda-web-server/src/index.ts
@@ -92,6 +92,11 @@ const argv = yargs
             "An older system to fall back to - this url will be shown in a banner that says 'you can still go back to old site'.",
         type: "string"
     })
+    .option("gapiIds", {
+        describe: "Google Analytics ID(s)",
+        type: "array",
+        default: []
+    })
     .option("datasetSearchSuggestionScoreThreshold", {
         describe:
             "The score threshold after which to display the dataset suggestion form",
@@ -172,7 +177,8 @@ const webServerConfig = {
     ),
     fallbackUrl: argv.fallbackUrl,
     datasetSearchSuggestionScoreThreshold:
-        argv.datasetSearchSuggestionScoreThreshold
+        argv.datasetSearchSuggestionScoreThreshold,
+    gapiIds: argv.gapiIds
 };
 
 app.get("/server-config.js", function(req, res) {